Deduplicate call log entries during restore
This change introduces a new flag, `call_log_restore_deduplication_enabled`, which enables deduplication of call log entries during restore operations. This prevents duplicate entries from being added to the call log when a user restores their device from a backup.

The deduplication logic checks if an entry with the same date and number already exists in the call log. If a duplicate is found, the entry is not restored.

Flag: com.android.calllogbackup.call_log_restore_deduplication_enabled
Bug: 371959886
Bug: 374931480
Test: Flash on device and check for duplicates after multiple runs of restore
Change-Id: I24e49a5629ea0083c460a50bee72439d31d22ed4